move notify-server to virt-handler
